% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_createColorVector.R
\name{createColorVector}
\alias{createColorVector}
\title{Function to create a color code vector for each subgroup}
\usage{
createColorVector(
  results,
  y,
  point_color,
  selected_subgroup,
  selected_subgroup_color,
  filter1,
  filter2,
  variableChosen1,
  variableChosen2,
  filter_color,
  parent_subgroups_color,
  factorial_context_color,
  importance_values,
  importance_values_color,
  brightness,
  memorized_data,
  memorized_color,
  key
)
}
\arguments{
\item{results}{results data set object of class "SubScreenResult".}

\item{y}{target variable name.}

\item{point_color}{color of circles.}

\item{selected_subgroup}{selected subgroups.}

\item{selected_subgroup_color}{selected subgroup color.}

\item{filter1}{variable name of first filter.}

\item{filter2}{variable name of second filter.}

\item{variableChosen1}{level of first filter variable.}

\item{variableChosen2}{level of second filter variable.}

\item{filter_color}{color for filtered circles.}

\item{parent_subgroups_color}{color for parent circles.}

\item{factorial_context_color}{color for factorial context.}

\item{importance_values}{subgroup ids with importance values.}

\item{importance_values_color}{importance values color.}

\item{brightness}{brightness of circles.}

\item{memorized_data}{subgroup ids which are memorized.}

\item{memorized_color}{memorized subgroup id color.}

\item{key}{factor levels.}
}
\description{
Function to create a color code vector for each subgroup
}
\keyword{internal}
