% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_createFilteredTable.R
\name{createFilteredTable}
\alias{createFilteredTable}
\title{Function to create the subgroup filter table}
\usage{
createFilteredTable(
  filter1,
  filter2,
  variableChosen1,
  variableChosen2,
  results,
  y,
  x,
  bg.color,
  key
)
}
\arguments{
\item{filter1}{variable name of first filter.}

\item{filter2}{variable name of second filter.}

\item{variableChosen1}{level of first filter variable.}

\item{variableChosen2}{level of second filter variable.}

\item{results}{results data set object of class "SubScreenResult".}

\item{y}{target variable name.}

\item{x}{variable name.}

\item{bg.color}{background color.}

\item{key}{number factors.}
}
\description{
Function to create the subgroup filter table
}
\keyword{internal}
