% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_createParentTable.R
\name{createParentTable}
\alias{createParentTable}
\title{Function to create the subgroup parent table}
\usage{
createParentTable(results, parents, y, x, x2, bg.color, navpanel)
}
\arguments{
\item{results}{results data set object of class "SubScreenResult".}

\item{parents}{subgroup ids parents.}

\item{y}{target variable name.}

\item{x}{variable name.}

\item{x2}{second variable name.}

\item{bg.color}{background color.}

\item{navpanel}{navpanel id ("SubscreenExplorer"/"SubscreenComparer").}
}
\description{
Function to create the subgroup parent table
}
\keyword{internal}
