% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_createPlot_points_data_complement.R
\name{createPlot_points_data_complement}
\alias{createPlot_points_data_complement}
\title{Function to create a data set with complement information based on selected subgroup}
\usage{
createPlot_points_data_complement(results_tmp, y, sel_ids)
}
\arguments{
\item{results_tmp}{subscreen data set}

\item{y}{target variable}

\item{sel_ids}{selected subgroup id}
}
\description{
Function to create a data set with complement information based on selected subgroup
}
\keyword{internal}
