% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_displayOptionsPanel.R
\name{displayOptionsPanel}
\alias{displayOptionsPanel}
\title{shiny widgets of display option panel}
\usage{
displayOptionsPanel()
}
\description{
shiny widgets of display option panel
}
\keyword{internal}
