% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subscreen_interactionPlot.R
\name{interaction_plot2}
\alias{interaction_plot2}
\title{Creates an interaction plot used in Explorer and ASMUS-tab in Subgroup Explorer}
\usage{
interaction_plot2(
  df_data,
  fac1,
  fac2 = NULL,
  fac3 = NULL,
  response,
  bg.col = "#6B6B6B",
  bg.col2 = NULL,
  font.col = "white",
  y.min = "NA",
  y.max = "NA",
  box.col = "white",
  sg_green = "#5cb85c",
  sg_blue = "#3a6791",
  plot_type = ""
)
}
\arguments{
\item{df_data}{data frame with factorial context}

\item{fac1}{name of factor level 1}

\item{fac2}{name of factor level 2 (default: NULL)}

\item{fac3}{name of factor level 3 (default: NULL)}

\item{response}{target variable}

\item{bg.col}{background color}

\item{bg.col2}{second background color}

\item{font.col}{font color}

\item{y.min}{y-axis mininum.}

\item{y.max}{y-axis maximum.}

\item{box.col}{box color.}

\item{sg_green}{hex code for color palette creation.}

\item{sg_blue}{hex code for color palette creation.}

\item{plot_type}{linear ("") or logarithmic ("log") y-axis (default: "").}
}
\description{
Creates an interaction plot used in Explorer and ASMUS-tab in Subgroup Explorer
}
\keyword{internal}
