% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subscreen_parents.R
\name{parents}
\alias{parents}
\title{Returns all 'parent'-subgroups of a specific subgroup}
\usage{
parents(data, SGID)
}
\arguments{
\item{data}{The "SubScreenResult" object generated via function 'subscreencalc'.}

\item{SGID}{Subgroup id(s) of the subgroup for which the 'parent'-subgroups are requested.}
}
\value{
List of 'parent'-subgroups.
}
\description{
Returns all 'parent'-subgroups of a specific subgroup
}
\keyword{internal}
