% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pseudo_contexts.R
\name{pseudo_contexts}
\alias{pseudo_contexts}
\title{Function for adding the status of a factorial context ("complete"/"incomplete" or "pseudo complete") to the
SubScreenResult object (used in subscreencalc if parameter 'factorial = TRUE').}
\usage{
pseudo_contexts(data, endpoint, factors)
}
\arguments{
\item{data}{The list entry 'sge' from the "SubScreenResult" object generated via function 'subscreencalc'.}

\item{endpoint}{The vector of target variable(s).}

\item{factors}{The list entry 'factors' from the "SubScreenResult" object generated via function 'subscreencalc'.}
}
\description{
Function for adding the status of a factorial context ("complete"/"incomplete" or "pseudo complete") to the
SubScreenResult object (used in subscreencalc if parameter 'factorial = TRUE').
}
\keyword{internal}
