% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pseudo_func.R
\name{pseudo_func}
\alias{pseudo_func}
\title{Generate variables for complete/incomplete/pseudo complete factorial context(s)}
\usage{
pseudo_func(results, endpoint, factors)
}
\arguments{
\item{results}{The list entry 'sge' from the "SubScreenResult" object generated via function 'subscreencalc'.}

\item{endpoint}{The vector of target variable(s).}

\item{factors}{The list entry 'factors' from the "SubScreenResult" object generated via function 'subscreencalc'.}
}
\description{
Generate variables for complete/incomplete/pseudo complete factorial context(s)
}
\keyword{internal}
