% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Example results data set without factorial context and complement calculations}
\description{
Example results data set without factorial context and complement calculations
}
\keyword{data}
\keyword{internal}
