% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_complement_true.R
\name{results_complement_true}
\alias{results_complement_true}
\title{Example results data set without factorial context and with complement calculations}
\description{
Example results data set without factorial context and with complement calculations
}
\keyword{data}
\keyword{internal}
