% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_factorial_true.R
\name{results_factorial_true}
\alias{results_factorial_true}
\title{Example results data set with factorial context and without complement calculations}
\description{
Example results data set with factorial context and without complement calculations
}
\keyword{data}
\keyword{internal}
