% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subscreen_mosaicPlot.R
\name{subscreen_mosaicPlot}
\alias{subscreen_mosaicPlot}
\title{Creates an mosaic plot used in Mosaic-tab in Subgroup Explorer}
\usage{
subscreen_mosaicPlot(
  res,
  mos.x,
  mos.y = NULL,
  mos.y2 = NULL,
  mos.z,
  col.bg = c("#424242"),
  col.txt = c("#ffffff"),
  colrange.z = c("#10384FFF", "#00BCFFFF", "white", "#89D329FF", "#004422FF"),
  scale = "lin"
)
}
\arguments{
\item{res}{results data set from subscreencalc}

\item{mos.x}{first endpoint variable}

\item{mos.y}{second endpoint variable (default:NULL)}

\item{mos.y2}{third endpoint variable (default: NULL)}

\item{mos.z}{reference variable (mosaic size)}

\item{col.bg}{background color (default: '#424242')}

\item{col.txt}{text color font (default: '#ffffff')}

\item{colrange.z}{three color scale for mosaic colors (default: c('#00BCFF','gray89','#89D329'))}

\item{scale}{scale of endpoint values linear or logarithmic (default: 'lin')}
}
\description{
Creates an mosaic plot used in Mosaic-tab in Subgroup Explorer
}
\keyword{internal}
