% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_variableOptionsPanel.R
\name{variableOptionsPanel}
\alias{variableOptionsPanel}
\title{shiny widgets of variable option panel}
\usage{
variableOptionsPanel()
}
\description{
shiny widgets of variable option panel
}
\keyword{internal}
