% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_displayOptionsPanel.R
\name{displayOptionsPanel}
\alias{displayOptionsPanel}
\title{shiny widgets of display option panel}
\usage{
displayOptionsPanel(
  custom_ref_line_at_start,
  custom_ref_line_value,
  favour_label_at_start,
  favour_direction
)
}
\description{
shiny widgets of display option panel
}
\keyword{internal}
