% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscreenfunnel.R
\name{subscreenfunnel}
\alias{subscreenfunnel}
\title{Calculation of funnel shape for the subgroups explorer}
\usage{
subscreenfunnel(
  data,
  H,
  eval_function,
  min_start = 2,
  n_support_points = 50,
  nperm = 200,
  alpha = 0.001,
  stratified = TRUE,
  treat = "treat",
  endpoints = NULL,
  verbose = TRUE,
  nkernel = 1,
  ...
)
}
\arguments{
\item{data}{data frame with study data}

\item{H}{results file from subscreencalc}

\item{eval_function}{eval function used in subscreencalc}

\item{min_start}{integer value for minimal subgroup size value for permutation}

\item{n_support_points}{integer value for number of supportive points}

\item{nperm}{integer value for number of permutations}

\item{alpha}{numerical value}

\item{stratified}{logical value (TRUE/FALSE) for stratification}

\item{treat}{character value of treatment variabale name}

\item{endpoints}{character vector of endpoints}

\item{verbose}{logical value to switch on/off output of computational information (defaults to TRUE)}

\item{nkernel}{integer value for number of kernel used}

\item{...}{further parameters which where outdated used for notes and errors.}
}
\value{
an object of type SubScreenResult of the form
}
\description{
Calculation of funnel shape for the subgroups explorer
}
\keyword{analysis}
\keyword{subgroup}
