% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscreenshow.R
\name{subscreenshow}
\alias{subscreenshow}
\title{(ii) Visualization}
\usage{
subscreenshow(
  scresults = NULL,
  variable_importance = NULL,
  host = NULL,
  port = NULL,
  NiceNumbers = c(1, 1.5, 2, 4, 5, 6, 8, 10),
  windowTitle = "Subgroup Explorer",
  graphSubtitle = NULL,
  favour_label_verum_name = NULL,
  favour_label_comparator_name = NULL,
  showTables = FALSE,
  reference_line_at_start = FALSE,
  reference_value = 1,
  favour_label_at_start = FALSE,
  favour_direction = TRUE,
  subgroup_levels_at_start = NA,
  yaxis_type = "lin",
  add_funnel_at_start = FALSE
)
}
\arguments{
\item{scresults}{SubScreenResult object with results from a subscreencalc call}

\item{variable_importance}{variable importance object calculated via subscreenvi to unlock
'variable importance'-tab in the app}

\item{host}{host name or IP address for Shiny display}

\item{port}{port number for Shiny display}

\item{NiceNumbers}{list of numbers used for a 'nice' scale}

\item{windowTitle}{title which is shown for the browser tab}

\item{graphSubtitle}{subtitle for explorer plot}

\item{favour_label_verum_name}{verum name for label use in explorer graph}

\item{favour_label_comparator_name}{comparator name for label use in explorer graph}

\item{showTables}{logical for display tables in 'Explorer tab'}

\item{reference_line_at_start}{logical for reference line appearance at start (TRUE/FALSE)}

\item{reference_value}{numeric value of horizontal reference line}

\item{favour_label_at_start}{logical for favour labels appearance at start (TRUE/FALSE)}

\item{favour_direction}{logical for favor label direction (TRUE/FALSE) where TRUE means
favour_label_verum_name is on top}

\item{subgroup_levels_at_start}{integer value for subgroup level slider at start}

\item{yaxis_type}{character ("lin"/"log") for yaxis type}

\item{add_funnel_at_start}{logical for funnel appearance at start (TRUE/FALSE)}
}
\description{
Start the Shiny based interactive visualization tool to show the subgroup results
generated by subscreencalc.
See and explore all subgroup results at one glance. Pick and chose a specific
subgroup, the level of combinations or a certain factor with its combinations.
Switch easily between different endpoint/target variables.
}
\keyword{analysis}
\keyword{subgroup}
\keyword{visualization}
